+++ src/lib/Components/Inbox/ActiveBids/ActiveBid.tsx
@@ -1,4 +1,4 @@
-import OpaqueImageView from "lib/Components/OpaqueImageView"
+import OpaqueImageView from "lib/Components/OpaqueImageView/OpaqueImageView"
 import colors from "lib/data/colors"
 import SwitchBoard from "lib/NativeModules/SwitchBoard"
 import React from "react"
