+++ src/lib/Scenes/City/Components/TabFairItemRow/index.tsx
@@ -1,5 +1,5 @@
 import { Box, Flex, Sans, space } from "@artsy/palette"
-import OpaqueImageView from "lib/Components/OpaqueImageView"
+import OpaqueImageView from "lib/Components/OpaqueImageView/OpaqueImageView"
 import SwitchBoard from "lib/NativeModules/SwitchBoard"
 import { Fair } from "lib/Scenes/Map/types"
 import React from "react"
