+++ src/lib/Scenes/Fair/Components/FairBoothPreview/Components/FairBoothPreviewHeader.tsx
@@ -1,5 +1,5 @@
 import { Box, Button, Flex, Sans, Serif, space } from "@artsy/palette"
-import OpaqueImageView from "lib/Components/OpaqueImageView"
+import OpaqueImageView from "lib/Components/OpaqueImageView/OpaqueImageView"
 import React from "react"
 import { Dimensions, TouchableOpacity } from "react-native"
 import styled from "styled-components/native"
