+++ src/setupJest.ts
@@ -72,7 +72,7 @@ console.error = (message?: any) => {
 
 mockedModule("./lib/Components/SwitchView.tsx", "SwitchView")
 mockedModule("./lib/Components/Spinner.tsx", "ARSpinner")
-mockedModule("./lib/Components/OpaqueImageView.tsx", "AROpaqueImageView")
+mockedModule("./lib/Components/OpaqueImageView/OpaqueImageView.tsx", "AROpaqueImageView")
 mockedModule("./lib/Components/ArtworkGrids/InfiniteScrollGrid.tsx", "ArtworksGrid")
 
 // Artist tests
